+++ packages/browser-wallet/src/popup/pages/AddAccount/i18n/en.ts
@@ -5,6 +5,8 @@ const t = {
     createIdentity: 'Create new identity',
     maxAccountsReached:
         'This identity has reached its account creation limit. To create more accounts you must use another Concordium identity.',
+    accountRecovered: 'The account that was attempted to be created already exists. It has been added to the wallet.',
+    goToAccount: 'Go to the account',
 };
 
 export default t;