+++ packages/browser-wallet/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   SendTransaction now validates that an account has sufficient funds before sending a transaction (requested though the api).
 -   Support for eID (Criipto) identity document types.
+-   When creating an account that already exist, it will now be added to the wallet.
 
 ### Changed